Testing non-implemented route handling
This should allow routes to be nil, if they are nil it will simply register the NotImplemented func, which returns an error with http.StatusNotImplemented. This is simply to make it easier for me to visualize all routes before actually implementing them, and avoiding creating an extra function somewhere in the actual project.